Migrate undo-related classes to separate files and expose them
