+++ GDEdit/GDEdit/Application/Editor/Editor.cs
@@ -1658,102 +1658,6 @@ namespace GDEdit.Application.Editor
 
         private static string GetAppropriateForm(int count, string thing) => $"{count} {thing}{(count != 1 ? "" : "s")}";
         #endregion
-
-        private class UndoRedoSystem
-        {
-            // I know, the naming is terrible on this one (hence the documentation), we need to find a better name
-            /// <summary>Determines whether multiple actions will be logged in the undo stack.</summary>
-            public bool MultipleActionToggle;
-
-            /// <summary>The temporarily stored undoable action that is to be registered upon completing an operation.</summary>
-            public UndoableAction TemporaryUndoableAction = new UndoableAction();
-            /// <summary>The stack that contains all actions that are to be undone.</summary>
-            public Stack<UndoableAction> UndoStack = new Stack<UndoableAction>();
-            /// <summary>The stack that contains all actions that are to be redone.</summary>
-            public Stack<UndoableAction> RedoStack = new Stack<UndoableAction>();
-
-            /// <summary>Registers all the actions.</summary>
-            /// <param name="description"></param>
-            public void RegisterActions(string description)
-            {
-                TemporaryUndoableAction.Description = description;
-                UndoStack.Push(TemporaryUndoableAction);
-                TemporaryUndoableAction = new UndoableAction();
-                RedoStack.Clear();
-            }
-            /// <summary>Adds a temporary action to the temporary action object and registers the undoable action if the multiple action toggle is <see langword="false"/>.</summary>
-            /// <param name="description">The description of the actions.</param>
-            /// <param name="action">The action. It must only perform the changes the action performs without invoking the respective events.</param>
-            /// <param name="undo">The inverse action. It must only perform the changes the inverse action performs without invoking the respective events.</param>
-            public void AddTemporaryAction(string description, Action action, Action undo)
-            {
-                TemporaryUndoableAction.Add(action, undo);
-                if (!MultipleActionToggle)
-                    RegisterActions(description);
-            }
-
-            /// <summary>Undoes a number of actions. If the specified count is greater than the available actions to undo, all actions are undone.</summary>
-            /// <param name="count">The number of actions to undo.</param>
-            public void Undo(int count = 1)
-            {
-                int actions = Min(count, UndoStack.Count);
-                for (int i = 0; i < actions; i++)
-                {
-                    var action = UndoStack.Pop();
-                    action.Undo();
-                    RedoStack.Push(action);
-                }
-            }
-            /// <summary>Redoes a number of actions. If the specified count is greater than the available actions to redo, all actions are redone.</summary>
-            /// <param name="count">The number of actions to redo.</param>
-            public void Redo(int count = 1)
-            {
-                int actions = Min(count, RedoStack.Count);
-                for (int i = 0; i < actions; i++)
-                {
-                    var action = RedoStack.Pop();
-                    action.Redo();
-                    UndoStack.Push(action);
-                }
-            }
-        }
-
-        /// <summary>Contains information about an action that can be undone.</summary>
-        private class UndoableAction
-        {
-            private readonly List<Action> undos = new List<Action>();
-            private readonly List<Action> actions = new List<Action>();
-
-            /// <summary>The description of the undoable action.</summary>
-            public string Description { get; set; }
-
-            public UndoableAction() { }
-            public UndoableAction(string description) => Description = description;
-
-            /// <summary>Adds an undoable action to the action list.</summary>
-            /// <param name="action">The action to add to the list.</param>
-            /// <param name="undo">The undo action to add to the list.</param>
-            public void Add(Action action, Action undo)
-            {
-                actions.Add(action);
-                undos.Add(undo);
-            }
-
-            /// <summary>Undoes all the actions in the list.</summary>
-            public void Undo()
-            {
-                foreach (var u in undos)
-                    u.Invoke();
-            }
-            /// <summary>Redoes all the actions in the list.</summary>
-            public void Redo()
-            {
-                foreach (var a in actions)
-                    a.Invoke();
-            }
-            /// <summary>Clears the action list.</summary>
-            public void Clear() => actions.Clear();
-        }
     }
 
     #region Delegates
